backend/pyproject.toml
[tool.poetry]
name = "uvp-interactive-web"
version = "0.1.0"
description = ""
authors = ["Gary Tyler <mail@garytyler.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["all"], version = "^0.52.0"}
fastapi-utils = "^0.2.1"
aioredis = "^1.3.1"
broadcaster = {path = "../../broadcaster"}
python-slugify = "^4.0.0"
databases = "^0.2.6"
tortoise-orm = "^0.16.4"
# tortoise-orm = {path="../../tortoise-orm"}
pydantic = "^1.4"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
black = {version = "^19.10b0", allow-prereleases = true}
mypy = "^0.770"
pre-commit = "^2.2.0"
pyflakes = {git = "https://github.com/pycqa/pyflakes", rev = "1911c20"}
pytest = "^5.4.1"
pytest-asyncio = "^0.10.0"
asgi-lifespan = "^1.0.0"
websockets = "^8.1"
isort = "^4.3.21"
async-asgi-testclient = "^1.4.4"
pytest-xprocess = "^0.13.1"

[build-system]
requires = ["poetry>=0.12"]
build-app = "poetry.masonry.api"
